requirements.txt
streamlit
pymupdf
sentence-transformers
faiss-cpu
openai
numpy
requests
transformers